refactor(accelerator): extract accelerator-core crate + rewire GUI onto it (P1)
