+++ .github/workflows/accelerator.yml
@@ -72,9 +72,14 @@ jobs:
       - uses: ./.github/actions/setup-accelerator
         with:
           rust-cache-key: accelerator-desktop-x86_64-unknown-linux-gnu
-          rust-workspaces: packages/accelerator/src-tauri -> target
+          rust-workspaces: |
+            packages/accelerator/src-tauri -> target
+            packages/accelerator/core -> target
       - run: cargo fmt --check
       - run: cargo clippy --all-targets -- -D warnings
+      # accelerator-core is standalone (compiled as a path-dep of src-tauri, but not lint-covered by it).
+      - run: cargo fmt --check --manifest-path ../core/Cargo.toml
+      - run: cargo clippy --manifest-path ../core/Cargo.toml --all-targets -- -D warnings
 
   test:
     name: Rust Tests
@@ -90,9 +95,13 @@ jobs:
       - uses: ./.github/actions/setup-accelerator
         with:
           rust-cache-key: accelerator-desktop-x86_64-unknown-linux-gnu
-          rust-workspaces: packages/accelerator/src-tauri -> target
+          rust-workspaces: |
+            packages/accelerator/src-tauri -> target
+            packages/accelerator/core -> target
           rust-components: ""
       - run: cargo test
+      # accelerator-core's own unit tests (113) — not exercised by `cargo test` in src-tauri.
+      - run: cargo test --manifest-path ../core/Cargo.toml
 
   lint:
     name: Lint
