+++ packages/accelerator/core/src/server/auth.rs
