+++ packages/accelerator/core/src/versions.rs
