+++ packages/accelerator/src-tauri/Cargo.toml
@@ -24,6 +24,7 @@ name = "aztec-accelerator"
 path = "src/main.rs"
 
 [dependencies]
+accelerator-core = { path = "../core" }
 tauri = { version = "2", features = ["tray-icon", "image-png"] }
 tauri-plugin-webdriver = { version = "0.2", optional = true }
 axum = "0.8"
